requirements.txt
numpy==1.20.3
seaborn==0.11.1
matplotlib==3.4.2
pandas==1.2.4
streamlit==0.82.0
Pillow==8.2.0
